fix: reorganize folder structure and update API configuration
- Move auth datasources from datasource/ to datasources/ folder
- Move auth repository from repository/ to repositories/ folder
- Move home repository from repository/ to repositories/ folder
- Organize Arduino/ESP32 scripts into feature-specific folders
- Add ping-pong heartbeat to UNO_bt_gamepad script
- Update API base URL to production endpoint
- Fix .ino file paths in all module pages
- Add test_upload.ps1 script for debugging uploads

This commit fixes the folder structure to match Clean Architecture
conventions and ensures all Arduino scripts are properly organized.
