+++ lib/di/service_locator.dart
@@ -25,10 +25,10 @@ import '../core/presentation/bloc/bluetooth/bluetooth_bloc.dart';
 import '../core/storage/secure_storage_service.dart';
 import '../core/ui/snackbar_service.dart';
 import '../core/domain/usecases/share_file_usecase.dart';
-import '../features/auth/data/datasource/auth_remote_datasource.dart';
-import '../features/auth/data/datasource/auth_token_local_datasource.dart';
-import '../features/auth/data/datasource/auth_user_local_datasource.dart';
-import '../features/auth/data/repository/auth_repository_impl.dart';
+import '../features/auth/data/datasources/auth_remote_datasource.dart';
+import '../features/auth/data/datasources/auth_token_local_datasource.dart';
+import '../features/auth/data/datasources/auth_user_local_datasource.dart';
+import '../features/auth/data/repositories/auth_repository_impl.dart';
 import '../features/auth/domain/repositories/auth_repository.dart';
 import '../features/auth/domain/usecases/change_password.dart';
 import '../features/auth/domain/usecases/check_session.dart';
@@ -60,7 +60,7 @@ import '../features/gamepad/domain/repositories/gamepad_repository.dart';
 import '../features/gamepad/presentation/bloc/gamepad_bloc.dart';
 import '../features/home/data/datasources/home_local_datasource_impl.dart';
 import '../features/home/data/datasources/home_remote_datesource.dart';
-import '../features/home/data/repository/home_repository_impl.dart';
+import '../features/home/data/repositories/home_repository_impl.dart';
 import '../features/home/domain/repositories/home_repository.dart';
 import '../features/home/domain/usecases/get_remote_home_menu.dart';
 import '../features/home/presentation/bloc/home_bloc.dart';
