+++ lib/features/light_control/presentation/pages/light_control_page.dart
@@ -17,7 +17,7 @@ class LightControlPage extends StatelessWidget {
     interfaceRoute: LightControlInterfacePage.routeName,
     image: 'assets/images/static/light_control/light_control.png',
     videoId: 'kJpdoBLSmHs',
-    inoFile: 'assets/files/esp32_bt_light.ino',
+    inoFile: 'assets/files/esp32_bt_light/esp32_bt_light.ino',
     instructions: [
       InstructionItem(
         title:
